+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/FormItemSpec.kt
@@ -42,6 +42,7 @@ object FormItemSpecSerializer :
             "au_becs_bsb_number" -> BsbSpec.serializer()
             "au_becs_account_number" -> AuBankAccountNumberSpec.serializer()
             "au_becs_mandate" -> AuBecsDebitMandateTextSpec.serializer()
+            "boleto_tax_id" -> BoletoTaxIdSpec.serializer()
             "country" -> CountrySpec.serializer()
             "selector" -> DropdownSpec.serializer()
             "email" -> EmailSpec.serializer()