+++ paymentsheet-example/src/androidTest/java/com/stripe/android/test/core/FieldPopulator.kt
@@ -12,6 +12,7 @@ import androidx.test.espresso.Espresso
 import com.stripe.android.test.core.ui.Selectors
 import com.stripe.android.ui.core.elements.AddressSpec
 import com.stripe.android.ui.core.elements.AuBankAccountNumberSpec
+import com.stripe.android.ui.core.elements.BoletoTaxIdSpec
 import com.stripe.android.ui.core.elements.BsbSpec
 import com.stripe.android.ui.core.elements.CardBillingSpec
 import com.stripe.android.ui.core.elements.CardDetailsSectionSpec
@@ -27,7 +28,8 @@ class FieldPopulator(
     private val selectors: Selectors,
     private val testParameters: TestParameters,
     private val populateCustomLpmFields: () -> Unit,
-    private val verifyCustomLpmFields: () -> Unit = {}
+    private val verifyCustomLpmFields: () -> Unit = {},
+    private val values: Values,
 ) {
     private val formSpec = testParameters.paymentMethod.formSpec
 
@@ -94,10 +96,11 @@ class FieldPopulator(
         val cardExpiration: String = "1230",
         val cardCvc: String = "321",
         val auBecsBsbNumber: String = "000000",
-        val auBecsAccountNumber: String = "000123456"
+        val auBecsAccountNumber: String = "000123456",
+        val boletoTaxId: String = "00000000000",
     )
 
-    private fun verifyPlatformLpmFields(values: Values = Values()) {
+    private fun verifyPlatformLpmFields() {
         formSpec.items.forEach {
             when (it) {
                 is BsbSpec -> {
@@ -129,15 +132,20 @@ class FieldPopulator(
                 }
                 is AddressSpec -> {
                     if (testParameters.billing == Billing.Off) {
-                        // TODO: This will not work when other countries are selected or defaulted
                         selectors.getLine1()
                             .assertContentDescriptionEquals(values.line1)
                         selectors.getCity()
                             .assertContentDescriptionEquals(values.city)
-                        selectors.getZip()
-                            .assertContentDescriptionEquals(values.zip)
-                        selectors.getState()
-                            .assertTextContains(values.state)
+
+                        validateZip()
+
+                        if (usesStateDropdown()) {
+                            selectors.getState()
+                                .assertTextContains(values.state)
+                        } else {
+                            selectors.getState()
+                                .assertContentDescriptionEquals(values.state)
+                        }
                     }
                 }
                 is CountrySpec -> {}
@@ -151,9 +159,7 @@ class FieldPopulator(
                 is KlarnaCountrySpec -> {}
                 is CardBillingSpec -> {
                     if (testParameters.billing == Billing.Off) {
-                        // TODO: This will not work when other countries are selected or defaulted
-                        selectors.getZip()
-                            .assertContentDescriptionEquals(values.zip)
+                        validateZip()
                     }
                 }
                 else -> {}
@@ -161,12 +167,18 @@ class FieldPopulator(
         }
     }
 
-    private fun populatePlatformLpmFields(values: Values = Values()) {
+    private fun populatePlatformLpmFields() {
         formSpec.items.forEach {
             when (it) {
                 is BsbSpec -> {
                     selectors.getAuBsb().apply {
                         performTextInput(values.auBecsBsbNumber)
+                    }
+                }
+                is BoletoTaxIdSpec -> {
+                    selectors.getBoletoTaxId().apply {
+                        performScrollTo()
+                        performTextInput(values.boletoTaxId)
                     }
                 }
                 is CardDetailsSectionSpec -> {
@@ -202,15 +214,11 @@ class FieldPopulator(
                 is KlarnaCountrySpec -> {}
                 is CardBillingSpec -> {
                     if (testParameters.billing == Billing.Off) {
-                        // TODO: This will not work when other countries are selected or defaulted
-                        selectors.getZip().apply {
-                            performTextInput(values.zip)
-                        }
+                        populateZip()
                     }
                 }
                 is AddressSpec -> {
                     if (testParameters.billing == Billing.Off) {
-                        // TODO: This will not work when other countries are selected or defaulted
                         selectors
                             .getLine1()
                             .performScrollTo()
@@ -223,18 +231,22 @@ class FieldPopulator(
                             .performClick()
                             .performTextInput(values.city)
 
-                        selectors
-                            .getZip()
-                            .performScrollTo()
-                            .performClick()
-                            .performTextInput(values.zip)
+                        populateZip()
 
-                        selectors
-                            .getState()
-                            .performScrollTo()
-                            .performClick()
+                        if (usesStateDropdown()) {
+                            selectors
+                                .getState()
+                                .performScrollTo()
+                                .performClick()
 
-                        selectors.selectState(values.state)
+                            selectors.selectState(values.state)
+                        } else {
+                            selectors
+                                .getState()
+                                .performScrollTo()
+                                .performClick()
+                                .performTextInput(values.state)
+                        }
                     }
                 }
                 is CountrySpec -> {}
@@ -243,5 +255,37 @@ class FieldPopulator(
                 else -> {}
             }
         }
+    }
+
+    private fun populateZip() {
+        if (usesZip()) {
+            selectors.getZip().apply {
+                performScrollTo()
+                performTextInput(values.zip)
+            }
+        } else {
+            selectors.getPostalCode().apply {
+                performScrollTo()
+                performTextInput(values.zip)
+            }
+        }
+    }
+
+    private fun validateZip() {
+        if (usesZip()) {
+            selectors.getZip()
+                .assertContentDescriptionEquals(values.zip)
+        } else {
+            selectors.getPostalCode()
+                .assertContentDescriptionEquals(values.zip)
+        }
+    }
+
+    private fun usesStateDropdown(): Boolean {
+        return testParameters.merchantCountryCode in setOf("US", "GB")
+    }
+
+    private fun usesZip(): Boolean {
+        return testParameters.merchantCountryCode in setOf("US", "GB")
     }
 }